requirements.txt
asgiref==3.7.2
Django==4.2.5
django-admin==2.0.2
django-bootstrap5==23.3
django-excel-response2==3.0.6
django-six==1.0.5
django-taggit==4.0.0
django-tinymce==3.6.1
excel-base==1.0.4
isoweek==1.3.3
python-dateutil==2.8.2
screen==1.0.1
six==1.16.0
sqlparse==0.4.4
TimeConvert==3.0.13
typing_extensions==4.7.1
tzlocal==5.0.1
xlwt==1.3.0
